feat(layout): improve wallet layout
