+++ packages/frontend/src/widgets/User/ui.ts
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+export const StyledAccount = styled.div`
+    position: relative;
+    z-index: 10;
+
+    .accountMenu {
+        position: absolute;
+        top: 50px;
+    }
+`;
